requirements.txt
bcrypt==3.1.4
alembic==1.0.3
Flask==1.0.2
flask_cors==3.0.7
flask-login==0.4.1
flask_sqlalchemy==2.3.2
flask_script==2.0.6
psycopg2-binary==2.7.6.1
simplejson==3.16.0
requests==2.20.1
pytest==4.0.0
pytest-flask==0.14.0

